Add PaperStripWidget to render 30-note music box strips
So far all it does is render a strip for a 30-note music box on the
screen; next it needs some controls and the ability to print these strips.
So far it snaps time to 200ms half-beats, which is perhaps not necessary:
holes can be punched anywhere, just not too close together.  Should rather
be sensitive to the declared tempo, too.